+++ src/cmem/ConfidenceValue/ConfidenceValue.stories.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { ComponentStory, ComponentMeta } from "@storybook/react";
+import { ConfidenceValue } from "../../../index";
+
+export default {
+    title: "cmem/ConfidenceValue",
+    component: ConfidenceValue,
+    argTypes: {
+    }
+
+} as ComponentMeta<typeof ConfidenceValue>;
+
+const TemplateIcons: ComponentStory<typeof ConfidenceValue> = (args) => (
+    <ConfidenceValue {...args} />
+);
+export const Default = TemplateIcons.bind({});
+Default.args = {
+    value: "0.5"
+};